scrapers/requirements.txt
# Core web scraping libraries
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Advanced scraping tools
firecrawl-py>=0.0.16
jina>=3.16.0

# Data processing
pandas>=2.1.0
python-dateutil>=2.8.2

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
retrying>=1.3.4

# Optional: For headless browser automation
undetected-chromedriver>=3.5.4

